refactor: establish sites/brc-analytics landing zone and migrate calendarview (#1541)
