+++ pages/calendar/index.tsx
@@ -1,7 +1,7 @@
 import { BRC_PAGE_META } from "@/common/meta/brc/constants";
 import { StyledPagesMain } from "@/components/Layout/components/Main/main.styles";
 import { config } from "@/config/config";
-import { CalendarView } from "@/views/CalendarView/calendarView";
+import { CalendarView } from "@brc/views/CalendarView/calendarView";
 import { GetStaticProps } from "next";
 import { JSX } from "react";
 import { ROUTES } from "../../routes/constants";
