+++ tsconfig.json
@@ -17,6 +17,7 @@
     "baseUrl": ".",
     "paths": {
       "@/*": ["app/*"],
+      "@brc/*": ["sites/brc-analytics/*"],
       "@repo/shared": ["packages/shared/index.ts"],
       "@repo/shared/*": ["packages/shared/*"],
       "@site-config/*": ["site-config/*"],
